refactor(admin): mutation수정, hooks수정, css수정, 변수 수정, type수정
