+++ apps/admin/src/components/fair/FairForm/FairForm.tsx
@@ -4,46 +4,20 @@ import { flex } from '@maru/utils';
 import { styled } from 'styled-components';
 import { IconClock, IconCalendar } from '@maru/icon';
 import FormInput from '@maru/ui/src/Input/FormInput';
-import { formatFairRequestBody } from '@/utils/functions/getRequestBody';
-import type { FairFormInput } from '@/utils/functions/getRequestBody';
 import type { FairType } from '@/types/fair/client';
-import { useCreateFairMutation } from '@/services/fair/mutations';
-
+import { useFairForm } from '@/components/fair/FairForm/fair.hooks';
 import { useFairFormStore } from '@/store/fair/fairType';
 
 const FairForm = () => {
-  const createFairMutation = useCreateFairMutation();
+  const [form] = useFairFormStore();
 
-  const [form, setForm] = useFairFormStore();
-
-  const handleChange = <K extends keyof FairFormInput>(
-    key: K,
-    value: FairFormInput[K]
-  ) => {
-    setForm((prev) => ({
-      ...prev,
-      [key]: value,
-    }));
-  };
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    handleChange(name as keyof FairFormInput, value);
-  };
-
-  const handleDateChange = (value: string) => {
-    const newStart = value + form.start.slice(8);
-    setForm((prev) => ({ ...prev, start: newStart }));
-  };
-
-  const handleTimeChange = (value: string) => {
-    const newStart = form.start.slice(0, 8) + value;
-    setForm((prev) => ({ ...prev, start: newStart }));
-  };
-
-  const handleSubmit = () => {
-    const body = formatFairRequestBody(form);
-    createFairMutation.mutate(body);
-  };
+  const {
+    handleChange,
+    handleInputChange,
+    handleDateChange,
+    handleTimeChange,
+    handleSubmit,
+  } = useFairForm();
 
   return (
     <StyledFairForm>
@@ -59,7 +33,6 @@ const FairForm = () => {
           ]}
         />
       </CreateFormSort>
-
       <CreateFormSort>
         <Text fontType="context">장소</Text>
         <FormInput
@@ -69,7 +42,6 @@ const FairForm = () => {
           onChange={handleInputChange}
         />
       </CreateFormSort>
-
       <CreateFormSort>
         <Text fontType="context">입학 설명회 날짜 (8자리)</Text>
         <InputWrapper>
@@ -84,7 +56,6 @@ const FairForm = () => {
           </InputIconWrapper>
         </InputWrapper>
       </CreateFormSort>
-
       <CreateFormSort>
         <Text fontType="context">시간 (4자리)</Text>
         <InputWrapper>
@@ -99,7 +70,6 @@ const FairForm = () => {
           </InputIconWrapper>
         </InputWrapper>
       </CreateFormSort>
-
       <CreateFormSort>
         <Text fontType="context">신청 기한 (8자리)</Text>
         <CreateInputSort>
@@ -117,7 +87,6 @@ const FairForm = () => {
           />
         </CreateInputSort>
       </CreateFormSort>
-
       <CreateFairButton onClick={handleSubmit}>
         <Text fontType="btn1">새로운 입학전형 설명회 생성하기</Text>
       </CreateFairButton>
@@ -138,11 +107,20 @@ const StyledFairForm = styled.div`
 `;
 
 const CreateFormSort = styled.div`
-  display: flex;
   ${flex({ flexDirection: 'column' })};
   gap: 8px;
   width: 100%;
   margin-bottom: 24px;
+`;
+
+const CreateInputSort = styled.div`
+  ${flex({ alignItems: 'center', justifyContent: 'space-between' })};
+  gap: 10px;
+
+  & > * {
+    flex: 1;
+    max-width: 100%;
+  }
 `;
 
 const CreateFairButton = styled.button`
@@ -151,13 +129,6 @@ const CreateFairButton = styled.button`
   border-radius: 6px;
   margin-top: 24px;
   padding: 22px;
-`;
-
-const CreateInputSort = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 10px;
 `;
 
 const InputWrapper = styled.div`
@@ -171,6 +142,6 @@ const InputIconWrapper = styled.div`
   right: 12px;
   transform: translateY(-50%);
   display: flex;
-  gap: 4px;
   align-items: center;
+  pointer-events: none;
 `;
